SCP-3208 Separated the ContractPreview from the Contract page view
The Contract page view had 2 top level UI elements, one to be shown in the Dashboard view (now separated in the ContractPreview) and one to show when a particular contract was opened.

The main reason for this split is to try to decouple the ContractExecution state from the two places that it is shown.