Remove `AccountsDiff` infra from `Util`
This paves the way for the next step, which is to many the Marlowe
implementation independent from a larger-specific concept of "money". It
wasn't clear to me if `AccountsDiff` should be defined in terms of the
_ledger_ concept of money or the (to be introduced) _Marlowe_ concept,
but then it turned out this code is not used at all? Can of course
re-introduce if it's needed after all, but then we need to decide which
concept of money is appropriate here.