Introduce ledger-independent `Money`
The Marlowe semantics was using `Value` from `Ledger`, but this fixes
the token type (it hardcodes a currency symbol/string combination). In
this commit we introduce a ledger-independent `Money` type (essentially
a `Map Token Integer`) and use this instead, translating to `Value` only
when interacting with the ledger at the periphery.