Update `marlowe-actus`
The repeated use of

```haskell
Value (Observation Token) Token
```

(as opposed to simply `Value Token`) is a bit frustrating, especially
since the only reason for Value `being` polymorphic is a limitation of
the Plutus compiler. I've not tried to address this for now, because we
probably want to think about some type synonyms anyway that hide the
`Token` parameter (and others).