Update `marlowe-cli`
I've not attempted to make this `Token` agnostic, with the exception of
the Merkleization support.

The CLI defines a few orphan instances for `ToJSON`/`FromJSON`. I've
added an instance here for `Money`; we could instead add the instance
where `Money` is defined, but by defining it specifically for `Money
Token` (rather than `Money t`), we can define it in terms of
`moneyToValue`/`moneyFromValue`, which means this doesn't change the
JSON format. (It would be possible of course to give an instance for
`Money t` _and_ for `Money Token` in `marlowe` using overlapping
instances, but overlapping instances is not a great thing to depend on.)