Give Extended Marlowe its own copy of more types
Specifically, do not reuse `Party` (and the `AccountId` synonym) and
`ChoiceId`. This paves the way for the next step, which is generalizing
core Marlowe over identity.

Extended Marlowe is the language most users deal with, so by not
touching that we minimize the impact that the generalization has
downstream.

This also introduces

```haskell
-- | Injection for types where this is total
class ToCore' a b where
  toCore' :: a -> b
```

to go alongside `ToCore`, to avoid having to deal with
actually-impossible errors. `Party` and `ChoiceId` have instances of
`ToCore'`, as well as the existing duplicated type `Payee`.