SCP-4364 Some upgrade work done from Alonzo to Babbage
Work has also been started here to get the non-PAB tests working on the
public networks. To facilitate this, changes have been made to
Language.Marlowe.CLI.Transaction buildFaucet (not buildFaucet') to allow
the use of any wallet addr/skey as a "faucet" to fund another address.

The `marlowe-cli util faucet` command now uses buildFaucet (not
buildFaucet' as it did before) and switches have been added for
specifying the source fund addr/skey.

This relates to (but does not yet close) SCP-4364 and SCP-4146