Temporarily disable non-Linux on CI pending cardano-world fix.
Some of our usage of cardano-world is not compatible with evaluating
for one system on an incompatible host, e.g. evaluating for darwin on
the hydra evaluator running Linux. This will require changes to
cardano-world to fix, in the mean time this issue is already present
on the target branch so this will allow us to continue with CI.

Revert this soon, when cardano-world is updated.