SCP-4439 Added a new test for `marlowe-cli run auto-execute`
This new test, simpleSelcoins, exercises a new run command,
auto-execute, which uses the new `Language.Marlowe.CLI.Transaction
selectCoins` function to perform automatic coin selection.

This test is just a copy of `marlowe-cli/examples/simple` with no other
modifications to the test logic.

New code was also written to more reliably clean up the test addresses and
return funds to the faucet address. This change was made in the source test,
`examples/simple`, as well.

The ReadMe.md documentation for both tests was updated.

Finally, the Readme.md for all `examples` tests was updated to more clearly
explain how to generate the documentation.