SCP-4674 Fixed bug in `findMinUtxo`.
The original version of this function did a cast to Marlowe data
(via `Core.fromChainDatum Core.MarloweV1`) before it did another
cast to `ScriptData`. If the `Chain.Datum` was not interpretable
as Marlowe data, then the datum was discarded from the `minUTxO`
computation, resulting in an underestimate of the true `minUTxO`.